Replace manual user/password with a DATABASE_URL
In general, a PostGres database requires a username and a password in
order to access it.

Because it is quite common for a computer to access a database it has
locally, unless other username or password is specified, a system
called "postgres-ident" will pull your local credentials and use them
to log you in.

If you specify a username and password in your database yml, you
override postgres-ident.

PostGres databases, like many programming languages, have components to
interact over Networks. You can use the standard
protocol_like_http_or_postgres//username:password@host_machine:PORT_NUMBER/resource_i_want_like_a_database to log into a database, just like you might log into a remote machine via FTP or an SMTP server, etc. To do this, there is a convention to add a variable to the ENV called DATABASE_URL. This environment variable will then OVERRIDE even a username and password specified in a database yml file. The machine will go to it first and look for that remote server, log in and use that database. In fact, you can even use the DATABASE_URL to log into your production server LOCALLY. From there, you could, for example, make a pg_database_backup of your production database.

Rails or any application interacts with the database, which is a
separate program, and needs a username and password for it. To log in it will obtain (or error, complaining) through
one of the three following methods:
1) postgres-ident local username/password
2) Database yml-specified username/password
3) Database URL pattern, including username/password