Re-order JS asset load for Cocoon
A Cocoon form tutorial explains the importance of correct ordering of the loading of these JS assets. It MIGHT be a reason (the reason?) why the Cocoon after-load stuff isn't working. Or IS it now working since I re-added the .link class?

Anyway, here is a site with more info about correct implementation of
Cocoon forms:

https://www.sitepoint.com/better-nested-attributes-in-rails-with-the-cocoon-gem/